Ultimas modificaciones en base a codigo, y soluciones de errores, antes de subirlo a render y conectarlo con el database

+++ src/main/java/com/upc/inclutechtrabajo_parcial/controller/ForoController.java
@@ -32,7 +32,7 @@ public class ForoController {
         return ResponseEntity.ok(foroService.registrar(foro));
     }
     @DeleteMapping("/eliminar/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         foroService.eliminar(id);
         return ResponseEntity.noContent().build();
     }
