+++ src/main/java/com/upc/inclutechtrabajo_parcial/controller/ForoQueSigueUsuarioController.java
@@ -30,7 +30,7 @@ public class ForoQueSigueUsuarioController {
         return ResponseEntity.ok(foroSeguidoService.registrar(seguimiento));
     }
     @DeleteMapping("/eliminar/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         foroSeguidoService.eliminar(id);
         return ResponseEntity.noContent().build();
     }
